Rename head to bootstrap and update includes
