+++ public/driver/dashboard.php
@@ -1,6 +1,6 @@
 <?php
 //public/driver/dashboard.php
-require_once '../../includes/head.php';
+require_once '../../includes/bootstrap.php';
 
 // Fehleranzeige aktivieren (nur für Debugging, in Produktion entfernen)
 ini_set('display_errors', 1);
