+++ public/driver/update_entry.php
@@ -1,5 +1,5 @@
 <?php
-require_once '../../includes/head.php'; // Datenbankverbindung
+require_once '../../includes/bootstrap.php'; // Datenbankverbindung
 
 // Fehleranzeige aktivieren (nur für Debugging, in Produktion entfernen)
 ini_set('display_errors', 1);
