+++ public/fahrer_bearbeiten.php
@@ -1,5 +1,5 @@
 <?php
-require_once '../includes/head.php';
+require_once '../includes/bootstrap.php';
 
 ini_set('display_errors', 1);
 ini_set('display_startup_errors', 1);
