+++ public/fahrzeug_bearbeiten.php
@@ -1,5 +1,5 @@
 <?php
-require_once '../includes/head.php';
+require_once '../includes/bootstrap.php';
 
 if (!isset($_GET['id']) || empty($_GET['id'])) {
     die('Fahrzeug-ID fehlt!');
