+++ public/head.php
@@ -1,9 +1,3 @@
 <?php
-ini_set('display_errors', 1);
-ini_set('display_startup_errors', 1);
-error_reporting(E_ALL);
-
-require_once '../includes/auth.php';
-require_once '../includes/db.php';
-require_once '../includes/secure.php';
+require_once '../includes/bootstrap.php';
 ?>
