+++ public/schulungsverwaltung.php
@@ -4,7 +4,7 @@ ini_set('display_errors', 1);
 ini_set('display_startup_errors', 1);
 error_reporting(E_ALL);
 
-require_once '../includes/head.php';
+require_once '../includes/bootstrap.php';
 require_once __DIR__ . '/versand.php';
 
 function checkUndVersendeEinladungen($termin, $maxEinladungen = 8){
